requirements.txt
fastapi
uvicorn
python-dotenv
openai
python-multipart
google-cloud-storage
google-cloud-firestore

ffmpeg-python
